+++ frontend/rust-lib/flowy-server/src/af_cloud/impls/search.rs
@@ -1,14 +1,20 @@
 use crate::af_cloud::AFServer;
+use crate::util::tanvity_local_search;
 use flowy_ai_pub::cloud::search_dto::{
   SearchDocumentResponseItem, SearchResult, SearchSummaryResult,
 };
 use flowy_error::FlowyError;
 use flowy_search_pub::cloud::SearchCloudService;
+use flowy_search_pub::tantivy_state::DocumentTantivyState;
 use lib_infra::async_trait::async_trait;
+use std::sync::Weak;
+use tokio::sync::RwLock;
+use tracing::trace;
 use uuid::Uuid;
 
 pub(crate) struct AFCloudSearchCloudServiceImpl<T> {
-  pub inner: T,
+  pub server: T,
+  pub state: Option<Weak<RwLock<DocumentTantivyState>>>,
 }
 
 const DEFAULT_PREVIEW: u32 = 80;
@@ -23,12 +29,20 @@ where
     workspace_id: &Uuid,
     query: String,
   ) -> Result<Vec<SearchDocumentResponseItem>, FlowyError> {
-    let client = self.inner.try_get_client()?;
+    let client = self.server.try_get_client()?;
     let result = client
       .search_documents(workspace_id, &query, 10, DEFAULT_PREVIEW, None)
       .await?;
 
-    Ok(result)
+    if !result.is_empty() {
+      return Ok(result);
+    }
+
+    trace!("[Search] Local AI search returned no results, falling back to local search");
+    let items = tanvity_local_search(&self.state, workspace_id, &query)
+      .await
+      .unwrap_or_default();
+    Ok(items)
   }
 
   async fn generate_search_summary(
@@ -37,7 +51,7 @@ where
     query: String,
     search_results: Vec<SearchResult>,
   ) -> Result<SearchSummaryResult, FlowyError> {
-    let client = self.inner.try_get_client()?;
+    let client = self.server.try_get_client()?;
     let result = client
       .generate_search_summary(workspace_id, &query, search_results)
       .await?;
