+++ frontend/rust-lib/flowy-server/src/local_server/impls/search.rs
@@ -1,13 +1,11 @@
 use crate::af_cloud::define::LoggedUser;
+use crate::util::tanvity_local_search;
 use flowy_ai::local_ai::controller::LocalAIController;
-use flowy_ai_pub::cloud::search_dto::{
-  SearchDocumentResponseItem, SearchResult, SearchSummaryResult,
-};
 use flowy_error::FlowyError;
-use flowy_search_pub::cloud::SearchCloudService;
-use flowy_search_pub::entities::TanvitySearchResponseItem;
+use flowy_search_pub::cloud::{
+  SearchCloudService, SearchDocumentResponseItem, SearchResult, SearchSummaryResult,
+};
 use flowy_search_pub::tantivy_state::DocumentTantivyState;
-use flowy_server_pub::search_dto::SearchContentType;
 use lib_infra::async_trait::async_trait;
 use std::sync::{Arc, Weak};
 use tokio::sync::RwLock;
@@ -21,46 +19,6 @@ pub struct LocalSearchServiceImpl {
   pub state: Option<Weak<RwLock<DocumentTantivyState>>>,
 }
 
-impl LocalSearchServiceImpl {
-  async fn local_search(
-    &self,
-    workspace_id: &Uuid,
-    query: &str,
-  ) -> Option<Vec<SearchDocumentResponseItem>> {
-    match self.state.as_ref().and_then(|v| v.upgrade()) {
-      None => {
-        trace!("[Search] tanvity state is None");
-        None
-      },
-      Some(state) => {
-        let results = state.read().await.search(workspace_id, query, None).ok()?;
-        let items = results
-          .into_iter()
-          .flat_map(|v| tanvity_document_to_search_document(*workspace_id, v))
-          .collect::<Vec<_>>();
-        Some(items)
-      },
-    }
-  }
-}
-
-fn tanvity_document_to_search_document(
-  workspace_id: Uuid,
-  doc: TanvitySearchResponseItem,
-) -> Option<SearchDocumentResponseItem> {
-  let object_id = Uuid::parse_str(&doc.id).ok()?;
-  Some(SearchDocumentResponseItem {
-    object_id,
-    workspace_id,
-    score: 1.0,
-    content_type: Some(SearchContentType::PlainText),
-    content: doc.content,
-    preview: None,
-    created_by: "".to_string(),
-    created_at: Default::default(),
-  })
-}
-
 #[async_trait]
 impl SearchCloudService for LocalSearchServiceImpl {
   async fn document_search(
@@ -68,23 +26,12 @@ impl SearchCloudService for LocalSearchServiceImpl {
     workspace_id: &Uuid,
     query: String,
   ) -> Result<Vec<SearchDocumentResponseItem>, FlowyError> {
+    let mut results = vec![];
     #[cfg(any(target_os = "windows", target_os = "macos", target_os = "linux"))]
     {
       if let Ok(scheduler) = flowy_ai::embeddings::context::EmbedContext::shared().get_scheduler() {
         match scheduler.search(workspace_id, &query).await {
-          Ok(results) => {
-            return if results.is_empty() {
-              trace!("[Search] Local AI search returned no results, falling back to local search");
-              let items = self
-                .local_search(workspace_id, &query)
-                .await
-                .unwrap_or_default();
-              trace!("[Search] Local search returned {} results", items.len());
-              Ok(items)
-            } else {
-              Ok(results)
-            }
-          },
+          Ok(items) => results = items,
           Err(err) => tracing::error!("[Search] Local AI search failed: {:?}", err),
         }
       } else {
@@ -92,7 +39,15 @@ impl SearchCloudService for LocalSearchServiceImpl {
       }
     }
 
-    Ok(Vec::new())
+    if !results.is_empty() {
+      return Ok(results);
+    }
+
+    trace!("[Search] Local AI search returned no results, falling back to local search");
+    let items = tanvity_local_search(&self.state, workspace_id, &query)
+      .await
+      .unwrap_or_default();
+    Ok(items)
   }
 
   async fn generate_search_summary(
